src/main/java/com/jaewoo/batch/app/basic/HelloWorldJob.java
package com.jaewoo.batch.app.basic;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.batch.core.Job;
import org.springframework.batch.core.Step;
import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
import org.springframework.batch.core.launch.support.RunIdIncrementer;
import org.springframework.batch.repeat.RepeatStatus;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;

@Configuration
@RequiredArgsConstructor
@Slf4j
public class HelloWorldJob {

    private final JobBuilderFactory jobBuilderFactory;

    private final StepBuilderFactory stepBuilderFactory;

    @Bean
    public Job helloWorldJob_batchBuilder() {

        return jobBuilderFactory.get("helloWorldJob")
                .incrementer(new RunIdIncrementer())
                .start(helloWorldJob_step01())
                .build();
    }

    @Bean
    public Step helloWorldJob_step01() {
        return stepBuilderFactory.get("helloWorldJob_step01")
                .tasklet((x, y) -> {
                    log.info("tasklet log Job -> Step");
                    return RepeatStatus.FINISHED;
                }).build();
    }
}
